Refactor graphML: extract helpers
