+++ src/main/utils/formatters/index.js
@@ -1,7 +1,7 @@
+const GraphMLFormatter = require('./graphml/GraphMLFormatter');
 const { AdjacencyMatrixFormatter } = require('./matrix');
 const { AdjacencyListFormatter } = require('./adjacency-list');
 const { AttributeListFormatter } = require('./attribute-list');
-const { GraphMLFormatter } = require('./graphml/createGraphML');
 
 module.exports = {
   AdjacencyMatrixFormatter,
